backend/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
pydantic
python-multipart
google-generativeai
pypdf
numpy
python-dotenv
